fix(compression): split RTK config schema
